// packages/lib/src/models/external.ts
export type SdkLayout = 'pip' | 'sbs';

export type SdkPrimaryContent = 'media' | 'slides';

export type SdkPipPosition = 'bottom-left' | 'bottom-right' | 'top-left' | 'top-right';

export interface SdkCaptionTrack {
  guid: string;
  languageCode: string;
  source: 'AUTOGENERATED' | 'CAPTIONHUB' | 'ENCODER' | 'UPLOADED';
  title: string;
}